Backed out changeset 18f02d991078 (bug 956899) for likely causing intermittent crashes in 13.3.0_7.js. r=backout
UltraBlame original commit: 764ab2ad75e784d0175f6645c1c2fca4816863af

+++ js/src/threading/windows/ConditionVariable.cpp
@@ -381,12 +381,10 @@ js::ConditionVariable::wait_for(UniqueLock<Mutex>& lock,
   CRITICAL_SECTION* cs = &lock.lock.platformData()->criticalSection;
 
   
-  
-  
   double msecd = rel_time.ToMilliseconds();
   DWORD msec = msecd < 0.0
                ? 0
-               : msecd > UINT32_MAX
+               : msecd > DBL_MAX
                  ? INFINITE
                  : static_cast<DWORD>(msecd);
